Remove redundant integer literal casts (N as c_int, N as lzma_vli, N as size_t)

+++ liblzma-rs/src/common/alone_encoder.rs
@@ -44,7 +44,7 @@ extern "C" fn write32le(buf: *mut u8, num: u32) {
         *buf.offset(3) = (num >> 24) as u8;
     }
 }
-pub const ALONE_HEADER_SIZE: c_int = 1 + 4 as c_int + 8 as c_int;
+pub const ALONE_HEADER_SIZE: c_int = 1 + 4 + 8;
 unsafe extern "C" fn alone_encode(
     coder_ptr: *mut c_void,
     allocator: *const lzma_allocator,
@@ -196,7 +196,7 @@ unsafe extern "C" fn alone_encoder_init(
     write32le((&raw mut (*coder).header as *mut u8).offset(1), d);
     memset(
         (&raw mut (*coder).header as *mut u8).offset(1).offset(4) as *mut c_void,
-        0xff as c_int,
+        0xff,
         8,
     );
     let filters: [lzma_filter_info; 2] = [
